seperate spectrum analyser constants to it's own header
namespace
